Update ftrace.js for large/complex systems
Hi all, I have a pretty complex set of contracts and I run into issues when tracing certain functions.

I've isolated the issue to ftrace.js and have been able to resolve the issue by modifying it slightly. Can share the codebase if anyone is interested in reproducing, but it is private at the moment.